requirements.txt
beautifulsoup4==4.10.0
certifi==2021.10.8
charset-normalizer==2.0.12
idna==3.3
lxml==4.8.0
numpy==1.19.5
Pillow==8.4.0
python-dateutil==2.8.2
python-pptx==0.6.21
pytz==2022.1
requests==2.27.1
six==1.16.0
soupsieve==2.3.1
urllib3==1.26.9
XlsxWriter==3.0.3
